Change `Spork` primitive to take `cont` and `spwn` arguments
For closure conversion, it is convenient to have arguments for the
`cont` and `spwn` functions available as SXML variables, rather than
needing to synthesize the `unit` argument.

Also, update closure conversion of `Spork` primitive to avoid a "no
closure convert info" property compiler error.  There is no need for
`newScope` with respect to the variable holding the result of `cont`,
because the applications of `seq` and `sync` to that variable occur in
the same SSA scope as the `contres` variable.

Aside: a `newScope` is for when the same SXML variable may be
translated to different SSA variables, depending on scope.  For
example, the free variables of a `Lambda`.  At the point where the
`Lambda` is evaluated (and the free variables captured in an
environment), the SXML variables are translated to one set of SSA
variables that are in scope.  But, in the defunctionalized SSA
function representing the `Lambda`'s code, the SXML variables will
correspond to a different set of SSA variables, namely the ones that
project the values of the environment record.

Also added wrappers to detect if the `cont` function raises and if the
`spwn` function raises or returns; similar wrappers were done for PCall.