Rework `spork` primitive to make data implicitly
Rework the `spork` higher-order primitive to have the type:

  ('aa -> 'ar) * 'aa *        (* `cont` function and argument*)
  ('ba * 'd -> 'br) * 'ba *   (* `spwn` function and argument, taking the argument and the join data *)
  ('ar -> 'c) *               (* `seq` continuation, taking the result of `cont` function *)
  ('ar * 'd -> 'c) ->         (* `sync` continuation, taking result of `cont` function and the join data *)
  'c

Now, there is no need for the `spork_getData` primitive (in SML source code).

At closureConversion, the `spork_getData` primitive is re-introduced.  This
ensures that the `spork_getData` primitives occur in the same SSA function as
the corresponding `Spork` and `Spoin` transfers.